Adds stoc T, standardized ylim, robust tpts
In the Kot equation solver:
Period T of the forcing can now be made a stochastic variable. The Y
limits for the X,Y, and Z plots are now set in code to be the same. This
was done for apples to apples comparisions between plots and variables.
Finally, any increasing array for tpts (the time points to solve the
equations for) should now run without error, and only these values will be
plotted in the results (formally, the initial conditions were always
plotted). This feature allows for suppressing transients.